Fixing ticket generation for each projection

+++ src/main/java/com/future/cinemaxx/repositories/TicketRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TicketRepo extends JpaRepository<Ticket, Integer> {
+public interface TicketRepo extends JpaRepository<Ticket, Long> {
 }
